+++ CHANGELOG.md
@@ -141,6 +141,7 @@
 - Fixed jump & land footstep sounds sometimes not playing when mixed with walking/running footsteps. They now use separate cooldowns.
 - Fixed gibs sometimes not creating blood particles, or spawning them inside the ground.
 - Fixed away a regional behavior difference in the config screen's ranged value element.
+- Ambience sound effects will now pause when the game window loses focus.
 
 # 5.0 BETA 14